fix(base-data): 修正教育资源类型字段的数据类型
- 将 getResourceType 方法的返回类型从 Long 修改为 Integer
- 移除不必要的 Long.valueOf 类型转换
- 确保方法返回值与字段定义类型一致

+++ src/main/java/com/insight/utils/pojo/basedata/Education.java
@@ -107,8 +107,8 @@ public class Education extends BaseXo {
         this.type = type;
     }
 
-    public Long getResourceType() {
-        return Long.valueOf(resourceType);
+    public Integer getResourceType() {
+        return resourceType;
     }
 
     public void setResourceType(Integer resourceType) {
